backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.3
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.1
whitenoise==6.7.0
